data/web/ControlRfid/requirements.txt
Django==1.9.2
djangorestframework
urllib3
pyTelegramBotAPI
django-admin-bootstrapped
django-extensions
django-bootstrap3
django_tables2
django-crispy-forms
django-graphos
ndg-httpsclient
PySocks
